renovate: automatically rebuild dist (#167)
Renovate is wrestling with pulumi-bot in
https://github.com/pulumi/verify-provider-release/pull/164 because
pulumi-bot is trying to re-build dist and Renovate wants to keep taking
ownership of the branch. This eats up unnecessary GHA minutes.

This PR adds some config so Renovate can re-build dist automatically on
its own.

`make renovate` is currently the only allowed post-run task, hence the
addition of a Makefile.

There's an alternative where we could configure Renovate to give up
ownership once pulumi-bot commits to a branch, but I'd like to move away
from workflows that involve coordinating between the two systems.